backend/requirements/dev.txt
flake8==3.9.2
pytest-django==4.4.0
pytest-env==0.6.2
pytest-asyncio==0.15.1
pytest-ordering==0.6
pytest-mock==3.6.1
freezegun==1.1.0
responses==0.13.4
watchdog==2.1.4
argh==0.26.2
black==21.7b0
pyinstrument==4.0.3
pyfakefs==4.5.0
pytest-xdist==2.3.0
responses==0.13.4
graphviz==0.17